First implementation of server side filtering
This is a best effort implementation of server side filtering of the all stream. The main limitation being due to the fact it’s the clients responsibility to determine if the the end of the stream has been reached by no events being returned from the server. This means that in a given slice if there are no events to return after filtering then the server will return the last unfiltered event.

The StringFilter class is used to generate a filter based upon a series of regex and exact matches upon the event type. This is then used in the all reader to filter out unwanted events.