Implementing ReadAllEventsForwardFiltered and ReadAllEventsBackwardFiltered
Implemented on both the server, client and http. Can filter by either stream id or event type using either a prefix or a regular expression. As well as the filter a "maxSearchWindow" can also be provided that  defines how many events should be searched before attempting to return the slice whether the required number of events are found or not.